update test for propagator with const-e vverlet
The test for propagator with constant-E and velocity-verlet algorithm is
added. The energy conservation is also added.

+++ include/propagator-set-common.h
@@ -21,7 +21,7 @@ namespace std {
   }
 
   void SetOutputInterval(UniqueParameter* GPrm, const void* poi) {
-    GPrm[OutputInterval]=*reinterpret_cast<const unsigned int*>(poi);
+    GPrm[OutputInterval]=*reinterpret_cast<const double*>(poi);
   }
 
   typedef void (*GSetFuncType4Propagator)(UniqueParameter*,const void*);
